+++ Cuda/Src/header/prime.h
@@ -17,8 +17,8 @@
 __global__ void eratosthene(int *list, uint64_t borne);
 __global__ void listNumbers(int *list);
 __global__ void copyTab(int *src, int *dest, int size);
-uint64_t primeList(int *list, int *result, uint64_t borne);
-uint64_t *generatePrimeList(uint64_t borne, int *size);
+int primeList(int *list, int *result, int borne);
+int *generatePrimeList(int borne, int *size);
 
 
 #endif /* PRIME_H_ */
