+++ Cuda/Src/header/smooth.h
@@ -5,6 +5,6 @@
 #include <stdlib.h>
 #include <stdint.h>
 
-int isBSmoothG(uint64_t *primeList, int size, uint64_t y);
+int isBSmoothG(int *primeList, int size, uint64_t y);
 
 #endif
