+++ Cuda/Src/smooth.cu
@@ -1,6 +1,6 @@
-#include "smooth.h"
+#include "header/smooth.h"
 
-int isBSmoothG(uint64_t *primeList, int size, uint64_t y){
+int isBSmoothG(int *primeList, int size, uint64_t y){
 	int i = 0;
 	if(i < size){
 		int y1 = y;
